extract showing of question to own component
